backend/requirements.txt
# FastAPI 核心依赖 (使用兼容 Python 3.13 的版本)
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# 数据库连接 (TaurusDB 兼容 MySQL)
PyMySQL==1.1.0
DBUtils==3.0.3  # 数据库连接池

# 安全与认证 (先安装 cryptography 预编译版本)
cryptography==41.0.7  # 支持 TLS 连接，使用预编译二进制包
python-jose==3.3.0  # JWT (不含 cryptography extra，避免重复安装)
passlib[bcrypt]==1.7.4  # 密码哈希
python-multipart==0.0.6  # 表单数据解析

# 工具库
python-dotenv==1.0.0  # 环境变量管理
loguru==0.7.2  # 日志

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # FastAPI 测试客户端

# API 文档增强（可选）
# swagger-ui-bundle==0.0.9
